IGNITE-13032 Use non-static stream pool for optimized marshaller. (#7814)
