+++ modules/core/src/main/java/org/apache/ignite/internal/marshaller/optimized/OptimizedObjectStreamRegistry.java
@@ -18,74 +18,22 @@
 package org.apache.ignite.internal.marshaller.optimized;
 
 import java.io.IOException;
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingQueue;
 import org.apache.ignite.IgniteException;
 import org.apache.ignite.internal.IgniteInterruptedCheckedException;
 import org.apache.ignite.internal.util.io.GridUnsafeDataInput;
 import org.apache.ignite.internal.util.io.GridUnsafeDataOutput;
-import org.apache.ignite.internal.util.typedef.internal.U;
 
 /**
  * Storage for object streams.
  */
-class OptimizedObjectStreamRegistry {
-    /** Holders. */
-    private static final ThreadLocal<StreamHolder> holders = new ThreadLocal<>();
-
-    /** Output streams pool. */
-    private static BlockingQueue<OptimizedObjectOutputStream> outPool;
-
-    /** Input streams pool. */
-    private static BlockingQueue<OptimizedObjectInputStream> inPool;
-
-    /**
-     * Ensures singleton.
-     */
-    private OptimizedObjectStreamRegistry() {
-        // No-op.
-    }
-
-    /**
-     * Sets streams pool size.
-     *
-     * @param size Streams pool size.
-     */
-    static void poolSize(int size) {
-        if (size > 0) {
-            outPool = new LinkedBlockingQueue<>(size);
-            inPool = new LinkedBlockingQueue<>(size);
-
-            for (int i = 0; i < size; i++) {
-                outPool.offer(createOut());
-                inPool.offer(createIn());
-            }
-        }
-        else {
-            outPool = null;
-            inPool = null;
-        }
-    }
-
+abstract class OptimizedObjectStreamRegistry {
     /**
      * Gets output stream.
      *
      * @return Object output stream.
      * @throws org.apache.ignite.internal.IgniteInterruptedCheckedException If thread is interrupted while trying to take holder from pool.
      */
-    static OptimizedObjectOutputStream out() throws IgniteInterruptedCheckedException {
-        if (outPool != null) {
-            try {
-                return outPool.take();
-            }
-            catch (InterruptedException e) {
-                throw new IgniteInterruptedCheckedException(
-                    "Failed to take output object stream from pool (thread interrupted).", e);
-            }
-        }
-        else
-            return holder().acquireOut();
-    }
+    abstract OptimizedObjectOutputStream out() throws IgniteInterruptedCheckedException;
 
     /**
      * Gets input stream.
@@ -93,83 +41,28 @@ class OptimizedObjectStreamRegistry {
      * @return Object input stream.
      * @throws org.apache.ignite.internal.IgniteInterruptedCheckedException If thread is interrupted while trying to take holder from pool.
      */
-    static OptimizedObjectInputStream in() throws IgniteInterruptedCheckedException {
-        if (inPool != null) {
-            try {
-                return inPool.take();
-            }
-            catch (InterruptedException e) {
-                throw new IgniteInterruptedCheckedException(
-                    "Failed to take input object stream from pool (thread interrupted).", e);
-            }
-        }
-        else
-            return holder().acquireIn();
-    }
+    abstract OptimizedObjectInputStream in() throws IgniteInterruptedCheckedException;
 
     /**
      * Closes and releases output stream.
      *
      * @param out Object output stream.
      */
-    static void closeOut(OptimizedObjectOutputStream out) {
-        U.close(out, null);
-
-        if (outPool != null) {
-            boolean b = outPool.offer(out);
-
-            assert b;
-        }
-        else {
-            StreamHolder holder = holders.get();
-
-            if (holder != null)
-                holder.releaseOut();
-        }
-    }
+    abstract void closeOut(OptimizedObjectOutputStream out);
 
     /**
      * Closes and releases input stream.
      *
      * @param in Object input stream.
      */
-    static void closeIn(OptimizedObjectInputStream in) {
-        U.close(in, null);
-
-        if (inPool != null) {
-            boolean b = inPool.offer(in);
-
-            assert b;
-        }
-        else {
-            StreamHolder holder = holders.get();
-
-            if (holder != null)
-                holder.releaseIn();
-        }
-    }
-
-    /**
-     * Gets holder from pool or thread local.
-     *
-     * @return Stream holder.
-     * @throws org.apache.ignite.internal.IgniteInterruptedCheckedException If thread is interrupted while trying to take holder from pool.
-     */
-    private static StreamHolder holder() throws IgniteInterruptedCheckedException {
-        StreamHolder holder = holders.get();
-
-        if (holder == null)
-            holders.set(holder = new StreamHolder());
-
-        return holder;
-    }
+    abstract void closeIn(OptimizedObjectInputStream in);
 
     /**
      * Creates output stream.
      *
      * @return Object output stream.
      */
-    private static OptimizedObjectOutputStream createOut() {
+    static OptimizedObjectOutputStream createOut() {
         try {
             return new OptimizedObjectOutputStream(new GridUnsafeDataOutput(4 * 1024));
         }
@@ -183,61 +76,12 @@ class OptimizedObjectStreamRegistry {
      *
      * @return Object input stream.
      */
-    private static OptimizedObjectInputStream createIn() {
+    static OptimizedObjectInputStream createIn() {
         try {
             return new OptimizedObjectInputStream(new GridUnsafeDataInput());
         }
         catch (IOException e) {
             throw new IgniteException("Failed to create object input stream.", e);
-        }
-    }
-
-    /**
-     * Streams holder.
-     */
-    private static class StreamHolder {
-        /** Output stream. */
-        private final OptimizedObjectOutputStream out = createOut();
-
-        /** Input stream. */
-        private final OptimizedObjectInputStream in = createIn();
-
-        /** Output streams counter. */
-        private int outAcquireCnt;
-
-        /** Input streams counter. */
-        private int inAcquireCnt;
-
-        /**
-         * Gets output stream.
-         *
-         * @return Object output stream.
-         */
-        OptimizedObjectOutputStream acquireOut() {
-            return outAcquireCnt++ > 0 ? createOut() : out;
-        }
-
-        /**
-         * Gets input stream.
-         *
-         * @return Object input stream.
-         */
-        OptimizedObjectInputStream acquireIn() {
-            return inAcquireCnt++ > 0 ? createIn() : in;
-        }
-
-        /**
-         * Releases output stream.
-         */
-        void releaseOut() {
-            outAcquireCnt--;
-        }
-
-        /**
-         * Releases input stream.
-         */
-        void releaseIn() {
-            inAcquireCnt--;
         }
     }
 }
